+++ src/main/java/com/sprint/hrbank/application/employee/dto/EmployeeCreateRequest.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 public record EmployeeCreateRequest(
     String name,
     String email,
-    Integer departmentId,
+    Long departmentId,
     String position,
     LocalDate hireDate,
     String memo) {}
